extract pathExists out to file-system module
