+++ app/src/lib/file-system.ts
@@ -17,3 +17,20 @@ export function mkdirIfNeeded(directoryPath: string): Promise<void> {
     })
   })
 }
+
+/**
+ * Helper function to promisify and simplify fs.stat.
+ *
+ * @param path Path to check for existence.
+ */
+export function pathExists(path: string): Promise<boolean> {
+  return new Promise<boolean>((resolve, reject) => {
+    Fs.stat(path, (error, stats) => {
+      if (error) {
+        resolve(false)
+      } else {
+        resolve(true)
+      }
+    })
+  })
+}
